Initial version. Execute: ./hircbot <host> <nick>
Currently it has the OnConnect event hardcoded to use the nick HIRCTEST and go to channel rochack. That will change.

It will just print out everything sent and received to standard output.

Uses runhaskell, which is on the CSUG network (I think?). You may have to run this before it Just Works: cabal install network